Split app into separate modules
